pyproject.toml
#
# https://packaging.python.org/en/latest/guides/writing-pyproject-toml/
#
# Minimum requirements for the build system to execute.
[build-system]
# PEP 508 specifications:
requires = ["setuptools", "wheel"]

# https://setuptools.readthedocs.io/en/latest/setuptools.html#configuring-setup-using-setup-cfg-files
[project]
name = "picframe"
version = "1.0.0"
description = "Apply a decorative frame to images"
readme = "README.rst"
authors = [
  {name = "Nathan Nichols", email = "nathannichols454@gmail.com"},
]
license = {file = "LICENSE"}
keywords = ["aesthetic", "design", "decoration"]
dependencies = [
  'Pillow >= 10.0.0',
  'numpy >= 1.0.0'
]

[project.scripts]
picframe = "picframe:beframe"

[project.urls]
Homepage = "https://resultsmotivated.com/"
Repository = "https://github.com/natelastname/picframe.git"
